Fix missing update in the tests folder

+++ tests/Fixtures/TestDriver.php
@@ -141,10 +141,10 @@ class TestDriver implements DriverInterface, VerifiesService
     /**
      * Send a typing indicator and wait for the given amount of seconds.
      * @param IncomingMessage $matchingMessage
-     * @param int $seconds
+     * @param float $seconds
      * @return mixed
      */
-    public function typesAndWaits(IncomingMessage $matchingMessage, int $seconds)
+    public function typesAndWaits(IncomingMessage $matchingMessage, float $seconds)
     {
         //
     }
